refactor(commands): extract /mcp command into mcp.rs
